+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 	- `ControllerHelperTrait::streamResponse()`
 	- `ControllerHelperTrait::jsonResponse()`
 * Removed the `$withAlpha` parameter from the `Color::toHexString()` method.
+* The `UploadedFile::moveTo()` method will now return a `FileInfo` instance instead of a boolean value.
 
 #### Improvements